requirements.txt
requests==2.32.4
python-dotenv==1.0.0
pytest==8.0.0
pytest-mock==3.12.0
ruff==0.3.0
backoff==2.2.1
